V1.Entrega Final - Funcionalidades OK

+++ src/sockets/socekt.manager.js
@@ -2,6 +2,10 @@ const socket = require("socket.io");
 const ProductRepository = require("../repositories/product.repository.js");
 const productRepository = new ProductRepository(); 
 const MessageModel = require("../models/message.models.js");
+const ProductManager = require("../controllers/product.manager.db.js");
+const productManager = new ProductManager
+const EmailManager = require("../services/email.js");
+const emailManager = new EmailManager();
 
 class SocketManager {
     constructor(httpServer) {
@@ -11,13 +15,28 @@ class SocketManager {
 
     async initSocketEvents() {
         this.io.on("connection", async (socket) => {
-            console.log("Nuevo cliente conectado.");
-            
+            console.log("Usuario conectado.");
+
             socket.emit("products", await productRepository.findAll() );
 
-            socket.on("deleteProduct", async (id) => {
-                await productRepository.deleteProduct(id);
-                this.emitUpdatedProducts(socket);
+            socket.on("deleteProduct", async ({ id, role }) => {
+                try {
+                    const product = await productRepository.findByID(id);
+                    if (!product) {
+                        socket.emit("deleteProductError", { error: "Producto no encontrado" });
+                        return;
+                    }
+                    await productRepository.deleteProduct(id);
+                    
+                    if (role === "admin" && product.owner !== "admin") {
+                        await emailManager.productDeleted(product.owner, product.title);
+                    }
+
+                    this.emitUpdatedProducts(socket);
+                } catch (error) {
+                    console.error(`Error al eliminar producto: ${error.message}`);
+                    socket.emit("deleteProductError", { error: "Error al eliminar producto" });
+                }
             });
 
             socket.on("createProduct", async (producto) => {
